# api/swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Duke decentralize bc",
    "version": "0.0.1",
    "title": "bc server"
  },
  "host": "localhost:3000",
  "basePath": "/api",
...
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json",
    "application/x-www-form-urlencoded",
    "application/xml",
    "text/xml"
  ],
  "produces": [
    "application/json",
    "application/xml",
    "text/xml",
    "application/javascript",
    "text/javascript"
  ],
  "paths": {
    "/Resident": {
      "get": {
        "tags": [
          "Resident"
        ],
        "summary": "Get all residents",
        "operationId": "getAllResidents",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Request was successful",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Resident"
              }
            }
          },
          "400": {
            "description": "Bad Request exception",
            "schema": {
              "$ref": "#/definitions/BadRequestError"
            }
          },
          "500": {
            "description": "Internal server exception",
            "schema": {
              "$ref": "#/definitions/InternalError"
            }
          }
        },
        "x-swagger-router-controller": "Resident"
      },
      "post": {
        "tags": [
          "Resident"
        ],
        "summary": "Create a new instance of the model and persist it into the data source.",
        "operationId": "createResident",
        "parameters": [
          {
            "in": "body",
            "name": "data",
            "description": "Model instance data",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Resident"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Request was successful",
            "schema": {
              "$ref": "#/definitions/Resident"
            }
          },
          "400": {
            "description": "Bad Request exception",
            "schema": {
              "$ref": "#/definitions/BadRequestError"
            }
          },
          "500": {
            "description": "Internal server exception",
            "schema": {
              "$ref": "#/definitions/InternalError"
            }
          }
        },
        "x-swagger-router-controller": "Resident"
      }
    },
    "/Resident/{id}": {
      "get": {
        "tags": [
          "Resident"
        ],
        "summary": "Find a resident by id",
        "operationId": "findResidentById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Resident's ID",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Request was successful",
            "schema": {
              "$ref": "#/definitions/Resident"
            }
          },
          "400": {
            "description": "Bad Request exception",
            "schema": {
              "$ref": "#/definitions/BadRequestError"
            }
          },
          "500": {
            "description": "Internal server exception",
            "schema": {
              "$ref": "#/definitions/InternalError"
            }
          }
        },
        "x-swagger-router-controller": "Resident"
      },
      "put": {
        "tags": [
          "Resident"
        ],
        "summary": "Update resident by ID.",
        "operationId": "updateResident",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Resident's ID",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "data",
            "description": "Resident's object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Resident"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Request was successful",
            "schema": {
              "$ref": "#/definitions/Resident"
            }
          },
          "400": {
            "description": "Bad Request exception",
            "schema": {
              "$ref": "#/definitions/BadRequestError"
            }
          },
          "500": {
            "description": "Internal server exception",
            "schema": {
              "$ref": "#/definitions/InternalError"
            }
          }
        },
        "x-swagger-router-controller": "Resident"
      },
      "delete": {
        "tags": [
          "Resident"
        ],
        "summary": "Delete resident by ID",
        "operationId": "deleteResidentById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Model id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Request was successful",
            "schema": {
              "type": "object",
              "properties": {}
            }
          },
          "400": {
            "description": "Bad Request exception",
            "schema": {
              "$ref": "#/definitions/BadRequestError"
            }
          },
          "500": {
            "description": "Internal server exception",
            "schema": {
              "$ref": "#/definitions/InternalError"
            }
          }
        },
        "x-swagger-router-controller": "Resident"
      }
    },
...
    "/Utility": {
      "get": {
        "tags": [
          "Utility"
        ],
        "summary": "Get all Utilities",
        "operationId": "getAllUtilities",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Request was successful",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Utility"
              }
            }
          },
          "400": {
            "description": "Bad Request exception",
            "schema": {
              "$ref": "#/definitions/BadRequestError"
            }
          },
          "500": {
            "description": "Internal server exception",
            "schema": {
              "$ref": "#/definitions/InternalError"
            }
          }
        },
        "x-swagger-router-controller": "Utility"
      },
      "post": {
        "tags": [
          "Utility"
        ],
        "summary": "Create a new Utility",
        "operationId": "createUtility",
        "parameters": [
          {
            "in": "body",
            "name": "data",
            "description": "Utility object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Utility"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Request was successful",
            "schema": {
              "$ref": "#/definitions/Utility"
            }
          },
          "400": {
            "description": "Bad Request exception",
            "schema": {
              "$ref": "#/definitions/BadRequestError"
            }
          },
          "500": {
            "description": "Internal server exception",
            "schema": {
              "$ref": "#/definitions/InternalError"
            }
          }
        },
        "x-swagger-router-controller": "Utility"
      }
    },
    "/Utility/{id}": {
      "get": {
        "tags": [
          "Utility"
        ],
        "summary": "Get Utility by id",
        "operationId": "findUtilityById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Utility id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Request was successful",
            "schema": {
              "$ref": "#/definitions/Utility"
            }
          },
          "400": {
            "description": "Bad Request exception",
            "schema": {
              "$ref": "#/definitions/BadRequestError"
            }
          },
          "500": {
            "description": "Internal server exception",
            "schema": {
              "$ref": "#/definitions/InternalError"
            }
          }
        },
        "x-swagger-router-controller": "Utility"
      },
      "put": {
        "tags": [
          "Utility"
        ],
        "summary": "Update Utility",
        "operationId": "updateUtility",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Utility id",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "data",
            "description": "Utility object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Utility"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Request was successful",
            "schema": {
              "$ref": "#/definitions/Utility"
            }
          },
          "400": {
            "description": "Bad Request exception",
            "schema": {
              "$ref": "#/definitions/BadRequestError"
            }
          },
          "500": {
            "description": "Internal server exception",
            "schema": {
              "$ref": "#/definitions/InternalError"
            }
          }
        },
        "x-swagger-router-controller": "Utility"
      },
      "delete": {
        "tags": [
          "Utility"
        ],
        "summary": "Delete Utility",
        "operationId": "deleteUtility",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Utility's ID",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Request was successful",
            "schema": {
              "type": "object",
              "properties": {}
            }
          },
          "400": {
            "description": "Bad Request exception",
            "schema": {
              "$ref": "#/definitions/BadRequestError"
            }
          },
          "500": {
            "description": "Internal server exception",
            "schema": {
              "$ref": "#/definitions/InternalError"
            }
          }
        },
        "x-swagger-router-controller": "Utility"
      }
    },
    "/Token": {
      "get": {
        "tags": [
          "Token"
        ],
        "summary": "Get all tokens",
        "operationId": "getAllTokens",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Request was successful",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Token"
              }
            }
          },
          "400": {
            "description": "Bad Request exception",
            "schema": {
              "$ref": "#/definitions/BadRequestError"
            }
          },
          "500": {
            "description": "Internal server exception",
            "schema": {
              "$ref": "#/definitions/InternalError"
            }
          }
        },
        "x-swagger-router-controller": "Token"
      },
      "post": {
        "tags": [
          "Token"
        ],
        "summary": "Create a new token",
        "operationId": "createToken",
        "parameters": [
          {
            "in": "body",
            "name": "data",
            "description": "Token object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Request was successful",
            "schema": {
              "$ref": "#/definitions/Token"
            }
          },
          "400": {
            "description": "Bad Request exception",
            "schema": {
              "$ref": "#/definitions/BadRequestError"
            }
          },
          "500": {
            "description": "Internal server exception",
            "schema": {
              "$ref": "#/definitions/InternalError"
            }
          }
        },
        "x-swagger-router-controller": "Token"
      }
    },
    "/Token/{id}": {
      "get": {
        "tags": [
          "Token"
        ],
        "summary": "Find a model instance by {{id}} from the data source.",
        "operationId": "findTokenById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Token's id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Request was successful",
            "schema": {
              "$ref": "#/definitions/Token"
            }
          },
          "400": {
            "description": "Bad Request exception",
            "schema": {
              "$ref": "#/definitions/BadRequestError"
            }
          },
          "500": {
            "description": "Internal server exception",
            "schema": {
              "$ref": "#/definitions/InternalError"
            }
          }
        },
        "x-swagger-router-controller": "Token"
      },
      "put": {
        "tags": [
          "Token"
        ],
        "summary": "Update token by id",
        "operationId": "updateToken",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Model id",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "data",
            "description": "Token's object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Request was successful",
            "schema": {
              "$ref": "#/definitions/Token"
            }
          },
          "400": {
            "description": "Bad Request exception",
            "schema": {
              "$ref": "#/definitions/BadRequestError"
            }
          },
          "500": {
            "description": "Internal server exception",
            "schema": {
              "$ref": "#/definitions/InternalError"
            }
          }
        },
        "x-swagger-router-controller": "Token"
      },
      "delete": {
        "tags": [
          "Token"
        ],
        "summary": "Delete token by id",
        "operationId": "deleteToken",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Token's id",
            "required": true,
            "type": "string",
            "format": "JSON"
          }
        ],
        "responses": {
          "200": {
            "description": "Request was successful",
            "schema": {
              "type": "object",
              "properties": {}
            }
          },
          "400": {
            "description": "Bad Request exception",
            "schema": {
              "$ref": "#/definitions/BadRequestError"
            }
          },
          "500": {
            "description": "Internal server exception",
            "schema": {
              "$ref": "#/definitions/InternalError"
            }
          }
        },
        "x-swagger-router-controller": "Token"
      }
    },
...
  },
  "definitions": {
    "InternalError": {
      "properties": {
        "message": {
          "type": "string",
          "default": "Internal Server Error"
        }
      },
      "description": "Internal Server Error"
    },
    "BadRequestError": {
      "properties": {
        "message": {
          "type": "string",
          "default": "Bad Request Error"
        }
      },
      "description": "Bad Request Error"
    },
    "Resident": {
      "type": "object",
      "required": [
        "alias",
        "cash",
        "energy",
        "firstName",
        "lastName",
        "tokens",
        "type"
      ],
      "properties": {
        "id": {
          "type": "string",
          "description": "Primary key for resident, backend creates uuid",
          "default": "sha256"
        },
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "alias": {
          "type": "string",
          "description": "Alias name to hide user"
        },
        "tokens": {
          "type": "string",
          "description": "The identifier of an struct of tokens"
        },
        "cash": {
          "type": "string",
          "description": "The identifier of an struct of cash"
        },
        "energy": {
          "type": "string",
          "description": "The identifier of an struct of energy"
        },
        "type": {
          "type": "string",
          "description": "Either Prosumer or Consumer",
          "enum": [
            "Prosumer",
            "Consumer"
          ]
        }
      },
      "description": "A participant named Resident",
      "example": {
        "firstName": "firstName",
        "lastName": "lastName",
        "alias": "alias",
        "tokens": "tokens",
        "id": "sha256",
        "type": "Prosumer",
        "cash": "cash",
        "energy": "energy"
      }
    },
...
    "Utility": {
      "type": "object",
      "required": [
        "energy",
        "id",
        "name",
        "tokens"
      ],
      "properties": {
        "id": {
          "type": "string",
          "description": "The instance identifier for struct UtilityCompany"
        },
        "name": {
          "type": "string"
        },
        "tokens": {
          "type": "string",
          "description": "The identifier of an struct of tokens"
        },
        "energy": {
          "type": "string",
          "description": "The identifier of an struct of energy"
        }
      },
      "description": "A participant named Utility",
      "example": {
        "name": "name",
        "tokens": "tokens",
        "id": "id",
        "energy": "energy"
      }
    },
    "Token": {
      "type": "object",
      "required": [
        "owner",
        "ownerType",
        "value"
      ],
      "properties": {
        "id": {
          "type": "string",
          "description": "The instance identifier for struct Tokens"
        },
        "value": {
          "type": "number",
          "format": "double"
        },
        "owner": {
          "type": "string"
        },
        "ownerType": {
          "type": "string",
          "enum": [
            "Resident",
            "Bank",
            "Utility"
          ]
        }
      },
      "description": "An asset named Tokens",
      "example": {
        "owner": "owner",
        "ownerType": "Resident",
        "id": "id",
        "value": 0.8008281904610115
      }
    },
...
  }
}